Move parse error check to CLI
